BUG: avoid sending HTTP GET requests with empty body
According to the HTTP specification [1], we should avoid sending a
body when doing a GET request.

In normal scenarios, sending an empty body doesn't cause any problems,
it simply gets ignored. However, some VPNs (like FortiNet) can be
configured to block these requests, as per the SPEC:

"[...] and might lead some implementations to reject the request
and close the connection because of its potential as a request
smuggling attack"

By avoiding these scenarios, we ensure we are spec-complient, and no
middleware software blocks the requests.

[1] - https://httpwg.org/specs/rfc9110.html#GET

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>